requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.7.0
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
Flask==3.1.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.10.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.0
requests==2.32.3
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
